# backend/openapi/course-paths.yaml
openapi: 3.0.4
info:
  title: Codillas Course API
  description: Course template — admin-authored courses with a Section → Lesson → Material structure.
  version: 0.3.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Codillas
    email: support@codillas.de
servers:
  - url: http://localhost:8081
    description: Local development
tags:
  - name: course
    description: Course catalog and Section → Lesson → Material structure authoring
paths:
  /api/courses:
    get:
      tags:
        - course
      operationId: listCourses
      summary: List courses (paginated)
      x-spring-paginated: true
      parameters:
        - $ref: "common.yaml#/components/parameters/PageNumber"
        - $ref: "common.yaml#/components/parameters/PageSize"
        - $ref: "common.yaml#/components/parameters/Sort"
        - name: status
          in: query
          required: false
          description: Optional lifecycle filter. Non-staff callers never receive DRAFT regardless.
          schema:
            $ref: "course-schemas.yaml#/components/schemas/CourseStatus"
      responses:
        "200":
          description: Paginated courses
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/CourseListResponse"
        "401":
          $ref: "common.yaml#/components/responses/Unauthorized"
    post:
      tags:
        - course
      operationId: createCourse
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "course-schemas.yaml#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/CourseResponse"
        "400":
          $ref: "common.yaml#/components/responses/BadRequest"
        "401":
          $ref: "common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
  /api/courses/{courseId}:
    get:
      tags:
        - course
      operationId: getCourse
      summary: Get a course with its full Section → Lesson structure
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course with nested sections and lesson summaries
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/CourseDetailResponse"
        "401":
          $ref: "common.yaml#/components/responses/Unauthorized"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
    delete:
      tags:
        - course
      operationId: deleteCourse
      summary: Delete a course and its entire Section → Lesson → Material structure
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
  /api/courses/{courseId}/publish:
    post:
      tags: [course]
      operationId: publishCourse
      summary: Publish a draft course (makes it selectable for groups)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/CourseResponse"
        "401": { $ref: "common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "common.yaml#/components/responses/NotFound" }
        "409": { $ref: "common.yaml#/components/responses/Conflict" }
  /api/courses/{courseId}/archive:
    post:
      tags: [course]
      operationId: archiveCourse
      summary: Archive a course (retire it; read-only thereafter)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/CourseResponse"
        "401": { $ref: "common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "common.yaml#/components/responses/NotFound" }
        "409": { $ref: "common.yaml#/components/responses/Conflict" }
  /api/courses/{courseId}/sections:
    post:
      tags:
        - course
      operationId: createSection
      summary: Add a section to a course
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "course-schemas.yaml#/components/schemas/CreateSectionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/SectionResponse"
        "400":
          $ref: "common.yaml#/components/responses/BadRequest"
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
  /api/sections/{sectionId}:
    patch:
      tags:
        - course
      operationId: updateSection
      summary: Rename a section
      parameters:
        - $ref: "#/components/parameters/SectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "course-schemas.yaml#/components/schemas/UpdateSectionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/SectionResponse"
        "400":
          $ref: "common.yaml#/components/responses/BadRequest"
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
    delete:
      tags:
        - course
      operationId: deleteSection
      summary: Delete a section and all its lessons and materials
      parameters:
        - $ref: "#/components/parameters/SectionId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
  /api/sections/{sectionId}/lessons:
    post:
      tags:
        - course
      operationId: createLesson
      summary: Add a lesson to a section
      parameters:
        - $ref: "#/components/parameters/SectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "course-schemas.yaml#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/LessonResponse"
        "400":
          $ref: "common.yaml#/components/responses/BadRequest"
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
  /api/lessons/{lessonId}:
    get:
      tags:
        - course
      operationId: getLesson
      summary: Get a lesson with its materials
      parameters:
        - $ref: "#/components/parameters/LessonId"
      responses:
        "200":
          description: Lesson with materials
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/LessonResponse"
        "401":
          $ref: "common.yaml#/components/responses/Unauthorized"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
    patch:
      tags:
        - course
      operationId: updateLesson
      summary: Update a lesson (title, summary, Meet link, recording URL)
      parameters:
        - $ref: "#/components/parameters/LessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "course-schemas.yaml#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/LessonResponse"
        "400":
          $ref: "common.yaml#/components/responses/BadRequest"
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
    delete:
      tags:
        - course
      operationId: deleteLesson
      summary: Delete a lesson and its materials
      parameters:
        - $ref: "#/components/parameters/LessonId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
  /api/lessons/{lessonId}/materials:
    post:
      tags:
        - course
      operationId: addMaterial
      summary: Attach a material (uploaded file or external link) to a lesson
      parameters:
        - $ref: "#/components/parameters/LessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "course-schemas.yaml#/components/schemas/CreateMaterialRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "course-schemas.yaml#/components/schemas/MaterialResponse"
        "400":
          $ref: "common.yaml#/components/responses/BadRequest"
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
  /api/materials/{materialId}:
    delete:
      tags:
        - course
      operationId: deleteMaterial
      summary: Delete a material
      parameters:
        - $ref: "#/components/parameters/MaterialId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "common.yaml#/components/responses/Forbidden"
        "404":
          $ref: "common.yaml#/components/responses/NotFound"
components:
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SectionId:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LessonId:
      name: lessonId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MaterialId:
      name: materialId
      in: path
      required: true
      schema:
        type: string
        format: uuid